pyproject.toml
[project]
name = "anyserve"
version = "0.1.0"
description = "anyServe - Capability-Oriented Serving Runtime"
authors = [
    { name = "simpx", email = "simpxx@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "requests>=2.31.0",
    "click>=8.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
]
llamacpp = [
    "llama-cpp-python>=0.2.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[project.scripts]
anyserve = "anyserve.cli:main"

[build-system]
requires = [
    "scikit-build-core>=0.5",
    "pybind11>=2.11",
]
build-backend = "scikit_build_core.build"

[tool.scikit-build]
cmake.source-dir = "cpp"
cmake.build-type = "Release"
cmake.args = [
    "-DBUILD_PYTHON_EXTENSION=ON",
    "-DCMAKE_TOOLCHAIN_FILE=build/conan_toolchain.cmake"
]
wheel.packages = ["python/anyserve"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "p0: Priority 0 (critical) tests",
    "p1: Priority 1 (high) tests",
    "p2: Priority 2 (medium) tests",
    "slow: Slow running tests",
    "integration: Integration tests",
]

[tool.coverage.run]
source = ["python/anyserve"]
omit = ["*/_proto/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
